refector distance join and knn join to be consistent with paper's names
