+++ engine/examples/src/main/scala/org/apache/spark/examples/sql/KNNJoinExample.scala
@@ -35,7 +35,7 @@ object KNNJoinExample {
     val sqlContext = new SQLContext(sc)
     sqlContext.setConf("spark.sql.shuffle.partitions", 16.toString)
     sqlContext.setConf("spark.sql.sampleRate", 1.toString)
-    sqlContext.setConf("spark.sql.joins.knnJoinMethod", "RTreeKNNJoin")
+    sqlContext.setConf("spark.sql.joins.KJSpark", "KJSpark")
 
     import sqlContext.implicits._
