+++ engine/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/CKJSpark.scala
@@ -29,11 +29,11 @@ import org.apache.spark.sql.spatial._
   * Created by dong on 1/20/16.
   * KNN Join based on Cartesian Product
   */
-case class CartesianKNNJoin(left_keys: Seq[Expression],
-                            right_keys: Seq[Expression],
-                            l: Literal,
-                            left: SparkPlan,
-                            right: SparkPlan) extends BinaryNode {
+case class CKJSpark(left_keys: Seq[Expression],
+                    right_keys: Seq[Expression],
+                    l: Literal,
+                    left: SparkPlan,
+                    right: SparkPlan) extends BinaryNode {
   override def outputPartitioning: Partitioning = left.outputPartitioning
 
   override def output: Seq[Attribute] = left.output ++ right.output
