+++ engine/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/NLKJSpark.scala
@@ -31,11 +31,11 @@ import scala.util.Random
   * Created by dong on 1/20/16.
   * KNN Join based on Nested Loop Approach
   */
-case class NestedLoopKNNJoin(left_keys: Seq[Expression],
-                             right_keys: Seq[Expression],
-                             l: Literal,
-                             left: SparkPlan,
-                             right: SparkPlan) extends BinaryNode {
+case class NLKJSpark(left_keys: Seq[Expression],
+                     right_keys: Seq[Expression],
+                     l: Literal,
+                     left: SparkPlan,
+                     right: SparkPlan) extends BinaryNode {
   override def output: Seq[Attribute] = left.output ++ right.output
 
   final val num_partitions = sqlContext.conf.numShufflePartitions
