+++ engine/sql/core/src/main/scala/org/apache/spark/sql/execution/joins/NLRDJSpark.scala
@@ -21,6 +21,7 @@ import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.util.NumberConverter
 import org.apache.spark.sql.execution.{BinaryNode, SparkPlan}
+import org.apache.spark.sql.index.RTree
 import org.apache.spark.sql.partitioner.MapDPartition
 import org.apache.spark.sql.spatial._
 
@@ -29,20 +30,18 @@ import scala.util.Random
 
 /**
   * Created by dong on 1/20/16.
-  * Distance Join based on Nested Loop Approach
+  * Distance Join based on Nested Loop + Local R-Tree
   */
-case class NestedLoopDistanceJoin(
-                                   left_keys: Seq[Expression],
-                                   right_keys: Seq[Expression],
-                                   l: Literal,
-                                   left: SparkPlan,
-                                   right: SparkPlan
-                                 ) extends BinaryNode {
+case class NLRDJSpark(left_keys: Seq[Expression],
+                      right_keys: Seq[Expression],
+                      l: Literal,
+                      left: SparkPlan,
+                      right: SparkPlan) extends BinaryNode {
   override def output: Seq[Attribute] = left.output ++ right.output
 
   final val num_partitions = sqlContext.conf.numShufflePartitions
   final val r = NumberConverter.literalToDouble(l)
-  final val dimension = left_keys.length
+  final val max_entries_per_node = sqlContext.conf.maxEntriesPerNode
 
   override protected def doExecute(): RDD[InternalRow] = {
     val tot_rdd = left.execute().map((0, _)).union(right.execute().map((1, _)))
@@ -69,26 +68,24 @@ case class NestedLoopDistanceJoin(
       while (iter.hasNext) {
         val data = iter.next()
         if (data._2._1 == 0) {
-          val tmp_point = new Point (left_keys.map(x =>
+          val tmp_point = new Point(left_keys.map(x =>
             BindReferences.bindReference(x, left.output).eval(data._2._2)
-            .asInstanceOf[Number].doubleValue()).toArray)
+              .asInstanceOf[Number].doubleValue()).toArray)
           left_data += ((tmp_point, data._2._2))
         } else {
-          val tmp_point = new Point (right_keys.map(x =>
+          val tmp_point = new Point(right_keys.map(x =>
             BindReferences.bindReference(x, right.output).eval(data._2._2)
-            .asInstanceOf[Number].doubleValue()).toArray)
+              .asInstanceOf[Number].doubleValue()).toArray)
           right_data += ((tmp_point, data._2._2))
         }
       }
 
       val joined_ans = mutable.ListBuffer[InternalRow]()
 
-      left_data.foreach {left =>
-        right_data.foreach {right =>
-          if (left._1.minDist(right._1) <= r) {
-            joined_ans += new JoinedRow(left._2, right._2)
-          }
-        }
+      if (right_data.nonEmpty) {
+        val right_rtree = RTree(right_data.map(_._1).zipWithIndex.toArray, max_entries_per_node)
+        left_data.foreach(left => right_rtree.circleRange(left._1, r)
+          .foreach(x => joined_ans += new JoinedRow(left._2, right_data(x._2)._2)))
       }
 
       joined_ans.iterator
